test(e2e): Wait for injector to appear
When jumping between proxied front-end apps (right now, this is the
"jump" between the log in app and the home page app), Jenkins has a hard
time rending fast enough to catch the DOM in a testable state.

https://github.com/angular/protractor/issues/1742#issuecomment-77457678

This was the summary of the problem. It appears that after logging in,
there is a "flash of white" that happens as you are taken to the next
angular app. At this point, an ng-app element appears, but no injector
has been set on it. Protractor detects this as an issue, when in reality
it just needs to wait a little while longer before it continues.

Your local machine is likely too slow to catch it, so it only happens on
Jenkins.

WARNING: this will happen again when we do full-integration testing.

There is a similar "flash of white" that occurs when moving from
Ticketing, Billing, or Servers and back. These types of situations will
be common in integration testing, so you should pay close attention when
moving between apps in those cases.